abstruct Merger
